+++ src/utils/url_utils.ts
@@ -11,8 +11,6 @@ export const generateAuthUrl = (url: string) => {
     const redirectUri = createRedirectUri();
 	_url.searchParams.append("source", VSCODE);
 	_url.searchParams.append("v", VERSION);
-	_url.searchParams.append("utm_medium", "plugin");
-	_url.searchParams.append("utm_source", VSCODE);
     _url.searchParams.append("redirect_uri", redirectUri);
 	return _url.href;
 };
@@ -27,10 +25,21 @@ export const generateServerUrl = (urlPath: string, baseUrl=API_BASE_URL, addTime
 	return _url.href;
 };
 
-export const generateWebUrl = (urlPath="", baseUrl=WEB_APP_URL) => {
-	const _url = new URL(`${baseUrl}${urlPath}`);
+export const generateWebUrl = (urlPath="", additionalParams: any=null) => {
+	const _url = appendGAparams(`${WEB_APP_URL}${urlPath}`);
+	if (additionalParams) {
+		Object.keys(additionalParams).forEach(key => {
+			_url.searchParams.append(key, additionalParams[key]);
+		});
+	}
+	return _url.href;
+};
+
+export const appendGAparams = (url: string) => {
+	console.log("appendGAparams", url);
+	const _url = new URL(url);
 	// Adding G4A params
 	_url.searchParams.append("utm_medium", "plugin");
 	_url.searchParams.append("utm_source", VSCODE);
-	return _url.href;
+	return _url;
 };