+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.45.0] - 2024-03-29
+- Using codesync.com for Auth instead of localhost browser
+- Upon Logout, closing Websocket connection so that no further diffs are processed
+
 ## [3.44.1] - 2024-03-26
 - Fixed Logout URL redirection